requirements.txt
# FastAPI and web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0

# Environment and configuration
python-dotenv==1.0.0

# Machine learning and similarity
numpy==1.26.3
scikit-learn==1.4.0
nltk==3.8.1

# LLM integration (UPDATED)
openai==1.10.0

# HTTP requests
requests==2.31.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4

# Load testing
locust==2.20.0